+++ crates/store/src/db/models/queries/nullifiers.rs
@@ -13,6 +13,7 @@ use diesel::{
 };
 use miden_lib::utils::{Deserializable, Serializable};
 use miden_node_utils::limiter::{
+    MAX_RESPONSE_PAYLOAD_BYTES,
     QueryParamLimiter,
     QueryParamNullifierLimit,
     QueryParamNullifierPrefixLimit,
@@ -21,7 +22,6 @@ use miden_objects::block::BlockNumber;
 use miden_objects::note::Nullifier;
 
 use super::DatabaseError;
-use crate::constants::MAX_PAYLOAD_BYTES;
 use crate::db::models::conv::{SqlTypeConvert, nullifier_prefix_to_raw_sql};
 use crate::db::models::utils::{get_nullifier_prefix, vec_raw_try_into};
 use crate::db::{NullifierInfo, schema};
@@ -69,7 +69,7 @@ pub(crate) fn select_nullifiers_by_prefix(
     pub const NULLIFIER_BYTES: usize = 32; // digest size (nullifier)
     pub const BLOCK_NUM_BYTES: usize = 4; // 32 bits per block number
     pub const ROW_OVERHEAD_BYTES: usize = NULLIFIER_BYTES + BLOCK_NUM_BYTES; // 36 bytes
-    pub const MAX_ROWS: usize = MAX_PAYLOAD_BYTES / ROW_OVERHEAD_BYTES;
+    pub const MAX_ROWS: usize = MAX_RESPONSE_PAYLOAD_BYTES / ROW_OVERHEAD_BYTES;
 
     assert_eq!(prefix_len, 16, "Only 16-bit prefixes are supported");