+++ crates/store/src/lib.rs
@@ -2,7 +2,6 @@ use std::time::Duration;
 
 mod accounts;
 mod blocks;
-mod constants;
 mod db;
 mod errors;
 pub mod genesis;